Added React to Print to project, made separate image react components hidden to be printed
Added react-to-print module to readily print react components at the click of a button.

I created 5 separate react components in order to use useRef's hook that react-to-print relies on. Each component has a separate useRef that is referenced whenever a print button is clicked. There is probably a better method of referencing separate simple components, but this is the work-around I found when messing with react-to-print.